\name{cvGEE}
\alias{cvGEE-package}
\alias{cvGEE}
\docType{package}

\title{
Proper Scoring Rules for Generalized Estimating Equations
}

\description{
Calculates the logarithmic, quadratic/Brier and spherical scoring rules based on generalized estimation equations.
}

\details{
\tabular{ll}{
Package: \tab cvGEE\cr
Type: \tab Package\cr
Version: \tab 0.3-5\cr
Date: \tab 2019-07-29\cr
License: \tab GPL (>=3)\cr
}

The package provides the estimated values of the scoring rules for each observation of the original dataset. These values can be summarized/averaged or used in figures to evaluate how the GEE performs in different ranges of the data.
}

\author{
Dimitris Rizopoulos

Maintainer: Dimitris Rizopoulos <d.rizopoulos@erasmusmc.nl>
}

\references{
Carvalho, A. (2016). An overview of applications of proper scoring rules. 
\emph{Decision Analysis} \bold{13}, 223-242. doi:10.1287/deca.2016.0337

Liang, K.Y. and Zeger, S.L. (1986). Longitudinal data  analysis  using  generalized  linear  models. \emph{Biometrika} \bold{73}, 13-22. doi:10.1093/biomet/73.1.13
}

\keyword{package}
\keyword{multivariate}

